requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Celery==2.4.3
Flask-HTTPAuth==2.2.1
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-RESTful==0.2.12
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-Uploads==0.1.3
Flask-WTF==0.10.0
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.3
PyJWT==0.2.1
SQLAlchemy==0.9.7
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
amqp==1.4.5
aniso8601==0.82
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.18
blinker==1.3
celery==3.1.13
certifi==14.05.14
decorator==3.4.0
flask-paginate==0.2.1
flower==0.7.3
httplib2==0.9
itsdangerous==0.24
kombu==3.0.21
oauthlib==0.6.3
pbr==0.10.0
pycrypto==2.6.1
python-gflags==2.0
python-openid==2.2.5
pytz==2014.4
redis==2.10.1
requests==2.3.0
requests-oauthlib==0.4.1
sinaweibopy==1.1.3
six==1.7.3
speaklater==1.3
wsgiref==0.1.2
ecdsa==0.11
qrcode==5.1
gevent
gevent-websocket==0.9.3
